arm64: Experimenting with adding a couple of callee save registers.
The idea is that this will allow the register allocator to keep a couple
values live over calls, reducing the need to spill in some cases. It also
has the benefit of increasing the number of registers available in general.

Currently has a bug relating to non-local exits. The debug code is
probably broken too.